Move api changeset close tests to their own module
